Connect compiler & simulator
Also a lot of bug fixes and QoL changes

+++ AST/Library.fs
@@ -2,7 +2,7 @@
 
 module CRNPP =
     type Species = string
-    type Number = float
+    type Number = decimal
 
     type NonComposableS = Cmp of Species * Species
